pyproject.toml
[project]
name = "sdn-ml-traffic-manager"
version = "0.1.0"
description = "Advanced ML-Driven Adaptive Traffic Management in SDN"
requires-python = ">=3.10"
dependencies = [
    "xgboost>=2.0.0",
    "scikit-learn>=1.3.0",
    "prometheus-client>=0.18.0",
    "pyyaml>=6.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "networkx>=3.1",
    "scipy>=1.11.0",
    "gymnasium>=0.29.0",
    "stable-baselines3>=2.0.0",
    "aiohttp>=3.9.0",
    "asyncio-mqtt>=0.13.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["sdn_ml"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.7.0",
    "mypy>=1.5.0",
]
